+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New parameter `overProvision`: when set available capacity on a node is calculated by taking into account
   the reserved capacity in the pool based on existing volumes.
 
+### Changed
+
+- When not using topology, select a random node to download the backup. This should prevent the same node
+  being used to download all backups.
+
 ## [1.2.3] - 2023-08-31
 
 ### Changed